RequestFactory uses HTTP_HOST instead of SERVER_PORT
The app environment should use the host and port that user's browser
actually requested.
The host and port the server is running on shouldn't affect the
functionality of the app, especially in environments where servers
are running behind reverse proxies etc where the SERVER_PORT could
be different to what the user facing app uses.